SPM: Added missing typecasts
This fixes following IAR warning:
    Error[Pe068]:
        integer conversion resulted in a change of sign

+++ secure_fw/spm/core/tfm_svcalls.c
@@ -157,14 +157,14 @@ static int32_t prepare_to_thread_mode_spm(uint8_t svc_number, uint32_t *ctx, uin
 
     if (svc_idx >= (sizeof(psa_api_svc_func_table)/sizeof(psa_api_svc_func_t))) {
         SPMLOG_ERRMSGVAL("Invalid PSA API SVC requested: ", svc_number);
-        ctx[0] = PSA_ERROR_GENERIC_ERROR;
+        ctx[0] = (uint32_t)PSA_ERROR_GENERIC_ERROR;
         return exc_return;
     }
 
     svc_func = psa_api_svc_func_table[svc_idx];
     if (!svc_func) {
         SPMLOG_ERRMSGVAL("Corresponding SVC function is not included for number ", svc_number);
-        ctx[0] = PSA_ERROR_GENERIC_ERROR;
+        ctx[0] = (uint32_t)PSA_ERROR_GENERIC_ERROR;
         return exc_return;
     }
 
@@ -181,7 +181,7 @@ static int32_t prepare_to_thread_mode_spm(uint8_t svc_number, uint32_t *ctx, uin
 
     init_spm_func_context(svc_func, ctx);
 
-    ctx[0] = PSA_SUCCESS;
+    ctx[0] = (uint32_t)PSA_SUCCESS;
 
     return EXC_RETURN_THREAD_PSP;
 }
@@ -241,7 +241,7 @@ static uint32_t handle_spm_svc_requests(uint32_t svc_number, uint32_t exc_return
 #endif
     default:
         SPMLOG_ERRMSGVAL("Unknown SPM SVC requested: ", svc_number);
-        svc_args[0] = PSA_ERROR_GENERIC_ERROR;
+        svc_args[0] = (uint32_t)PSA_ERROR_GENERIC_ERROR;
     }
 
     return exc_return;
@@ -306,7 +306,7 @@ uint32_t spm_svc_handler(uint32_t *msp, uint32_t exc_return, uint32_t *psp)
 #endif
 
     SPMLOG_ERRMSGVAL("Unknown SVC number requested: ", svc_number);
-    svc_args[0] = PSA_ERROR_GENERIC_ERROR;
+    svc_args[0] = (uint32_t)PSA_ERROR_GENERIC_ERROR;
 
     return exc_return;
 }
